Execute a command

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,7 +1,7 @@
 # manifest that kills a process named killmenow.
-
-exec { 'pkill':
-  command   => 'pkill -f killmenow',
-  provider  => 'Shell',
+exec { 'killmenow':
+  command     => 'pkill -f killmenow',
+  path        => '/bin:/usr/bin',
+  refreshonly => true,
+  onlyif      => 'pgrep -f killmenow',
 }
-
